% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.custom_timeseries_month_label}
\alias{.custom_timeseries_month_label}
\title{Creates some custom month/year labels for date axes}
\usage{
.custom_timeseries_month_label(x)
}
\description{
Creates some custom month/year labels for date axes
}
\keyword{internal}
