% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.term_trend_palette}
\alias{.term_trend_palette}
\title{Default palette applied to term trend lines in plotting functions}
\usage{
.term_trend_palette()
}
\description{
Default palette applied to term trend lines in plotting functions
}
\keyword{internal}
