% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-analysis.R
\name{.term_trends}
\alias{.term_trends}
\title{Fit a linear regression to time frequency series of term shares}
\usage{
.term_trends(termFrequencies, trendThreshold = 5e-04)
}
\arguments{
\item{termFrequencies}{a dataframe of \code{term} frequencies as returned by
\code{term_frequencies()}}

\item{trendThreshold}{a double used to categorise trends, default is
\code{0.0005}; if the intercept of the linear model of term frequencies
falls within a range of \code{+/-trendThreshold} of the mean term share,
the trend is categorized as \emph{"constant"}}
}
\value{
a dataframe specifying trend metrics for each \code{term}, where:
\describe{ \item{term}{a unique term} \item{slope}{the slope
coefficient of the linear model fit to this \code{term}'s time frequency
series} \item{volatility}{the residual standard deviation of the linear
model fit to this \code{term}'s time frequency series, which is used as a
basic measure of volatility of term frequencies} \item{trend}{a
categorisation of the term frequency trend, negative slopes with an
intercept \code{> mean term share + trendThreshold} are interpreted as
\emph{decreasing}, positive slopes with an intercept \code{< mean term
  share - trendThreshold} as \emph{increasing}, all others as
\emph{constant}} }
}
\description{
\code{.term_trends} takes a dataframe with term frequency time series as
returned by the \code{term_frequencies()} function and fits a linear
regression for each \code{term}.
}
\details{
The primary purpose of this function is to provide a simple approach to
identify \code{term} trends, which can be used for visualization, filtering
and exploratory analysis. Internally, the slope and intercept of the fitted
linear model are use to categorize a basic trend for the term frequencies.
}
\keyword{internal}
