% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-extract.R
\name{extract-term-ngrams}
\alias{extract-term-ngrams}
\alias{terms_by_date}
\alias{unigrams_by_date}
\alias{bigrams_by_date}
\title{Split a text source into tokens and terms by date of occurrence}
\usage{
terms_by_date(textData, textColumn, dateColumn, removeNumbers = TRUE,
  wordStemming = TRUE, customStopwords = NULL, tokenType = "unigram")

unigrams_by_date(textData, textColumn, dateColumn, removeNumbers = TRUE,
  wordStemming = TRUE, customStopwords = NULL)

bigrams_by_date(textData, textColumn, dateColumn, removeNumbers = TRUE,
  wordStemming = TRUE, customStopwords = NULL)
}
\arguments{
\item{textData}{a dataframe containing the text to be processed}

\item{textColumn}{a character string specifying the column name in
\code{textData} containing the text to be processed}

\item{dateColumn}{a character string specifying the column name in
\code{textData} specifying a publication date for the text in
\code{textColumn}}

\item{removeNumbers}{a Boolean indicating whether numbers should be removed
from the result; default is \strong{TRUE}.}

\item{wordStemming}{a Boolean indicating whether words in the text should be
reduced to the word stem; default is \strong{TRUE}.}

\item{customStopwords}{a character vector specifying additional stopwords
that should be removed from the result}

\item{tokenType}{the length of the consecutive token sequence extracted,
currently only \code{bigram} (two word sequence) and \code{unigram} (single
words) are supported, with \strong{\code{unigram} as default}}
}
\value{
a dataframe with three columns listing all individual term
occurrences in the provided text source, where \code{occur} is the
publication date associated with an original \code{token}, which has been
processed/reduced to \code{term}; if no stemming has been applied the term
and token in the result are identical
}
\description{
These functions transform a text source into a dataframe of individual terms
and tokens with an occurrence date. These terms/tokens can be extracted as
\emph{ngrams} of specified length. \code{terms_by_date} is wrapper around the
function for specific types of \emph{ngrams}.
}
\details{
Text input (\code{textColumn}) is split with a \emph{word tokenizer}, default
stopwords (see \code{\link[tidytext:stop_words]{tidytext}}) are removed and
tokens are further processed and filtered according to the function's
options. A \emph{term} is the character sequence obtained after all NLP
processing options this function offers have been applied, most importantly
\emph{stemming}, here the Porter stemmer from the
\code{\link[SnowballC:wordStem]{SnowballC package}} is applied.
}
