% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Convenience import of the pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
This makes the pipe operator available for all package functions.
}
\details{
See \code{\link[magrittr]{pipe}} for details.
}
\keyword{internal}
