% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-topics.R
\name{plot_term_frequencies}
\alias{plot_term_frequencies}
\title{Plot individual term frequencies by date in a faceted plot}
\usage{
plot_term_frequencies(termsByDate, timeBinUnit = "week",
  minTermTimeBins = 0.5, minTermOccurences = 10, topN = 25,
  selectTerms = NULL, selectTermsBy = "most_frequent",
  verboseLabels = FALSE, nCols = 5)
}
\arguments{
\item{termsByDate}{a dataframe as returned by \code{\link{terms_by_date}}}

\item{timeBinUnit}{a character sequence specifying the time period that
should be used as a bin unit when computing term frequencies. Valid values
are \code{"day", "week", "month", "quarter", "year"}, but for the text
sources processed in this package \code{"week"} is recommended and used as
a default. \strong{NOTE:} for the assignment of \code{week}s Monday is
considered as the first day of the week.}

\item{minTermTimeBins}{a double in the range \code{[0,1]} specifying the
minimum share of all unique timebins in which an occurrence of a term must
have been recorded, i.e. a value of \code{0.5} (the default) requires that
an occurrence of a term must have been recorded in at least 50\% of all
unique timebins covered by the dataset; terms that do not meet this
threshold will not be included in the returned results.}

\item{minTermOccurences}{an integer specifying the minimum of total
occurrences of a term to be included in the results; terms that do not meet
this threshold will not be included in the returned results.}

\item{topN}{the number of displayed top terms meeting the selection criteria
in \code{selectBy}}

\item{selectTerms}{a character vector of term patterns, that terms are
matched to for selection. \code{regular expression} syntax can be applied,
e.g. if \code{c("^mod", "an", "el$", "^outbreak$")} is supplied for
\code{selectTerms}, all terms that \strong{either} start with \emph{'mod'}
\strong{or} contain \emph{'an'} \strong{or} end with \emph{'el'} or the
exact term \emph{'outbreak'} are matched. The arguments \code{selectBy} and
\code{selectTerms} can be combined.}

\item{selectTermsBy}{the selection approach which determines the metric by which
\code{term}s will be sorted to select the \code{topN} terms. Currently, the
following options are supported: \describe{
\item{most_frequent}{\strong{the default}, select terms based on the total
number of occurrences} \item{trending_up}{select terms with largest upwards
trend; internally this is measured by the slope of a simple linear
regression fit to a \code{term}'s frequency series.}
\item{trending_down}{select terms with largest downward trend; internally
this is measured by the slope of a simple linear regression fit to a
\code{term}'s frequency series.} \item{trending}{select terms with either
largest upward or downward trend; internally this is measured by the
absolute value of the slope of a simple linear regression fit to a
\code{term}s frequency series.} \item{most_volatile}{select terms with the
largest change throughout the covered time period; internally this is
measured by the residual standard deviation of the linear model fit to a
\code{term}'s time frequency series.} }}

\item{verboseLabels}{a Boolean indicating if a single terms should be used as
labels for subplots or if multiple term instances (i.e. variations of the
term found in the original text source) should be used as labels. The
\strong{default is \code{FALSE}}.}

\item{nCols}{the number of columns along which term subplots should be layed
out.}
}
\description{
\code{plot_term_frequencies} plots time series of term shares for a selection
of terms in a faceted plot. Each term is displayed in a single subplot and
each time series is overlayed with a linear trendline.
}
\details{
This function merges the computation of term frequencies
(\code{\link{term_frequencies}}), creation of suitable labels
(\code{\link{terms_tokens_map}}) and the selection of terms by
msicalleneous criteria (\code{\link{select_top_terms}}) into one step.
}
\concept{visualizations}
