% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-topics.R
\name{plot_topic_frequencies}
\alias{plot_topic_frequencies}
\title{Plot individual topic frequencies by date in a faceted plot}
\usage{
plot_topic_frequencies(topicsByDocDate, topicLabels = NULL,
  timeBinUnit = "week", topN = 25, minTopicTimeBins = 0.5,
  minGamma = 0.01, selectTopicsBy = "most_frequent",
  selectTopics = NULL, verboseLabels = FALSE, nCols = 5)
}
\arguments{
\item{topicsByDocDate}{a dataframe as returned by
\code{\link{topics_by_doc_date}}}

\item{topicLabels}{a dataframe as returned by \code{\link{topics_terms_map}},
associating a \code{topic_id} with a suitable \code{topic_label}; if
\code{NULL} (the default), suitable default labels will be generated.}

\item{timeBinUnit}{a character sequence specifying the time period that
should be used as a bin unit when computing topic share frequencies. Valid
values are \code{"day", "week", "month", "quarter", "year"}, \code{"week"}
is the default. \strong{NOTE}, for the assignment of \code{week}s Monday is
considered as the first day of the week.}

\item{topN}{the number of top topics (according to the selection criteria in
\code{selectTopicsBy}) that should be displayed}

\item{minTopicTimeBins}{a double in the range \code{[0,1]} specifying the
minimum share of all unique timebins in which an occurrence of a topic
share of at least \code{minGamma} must have been recorded, i.e. a value of
\code{0.5} (the default) requires that an occurrence of a topic must have
been recorded in at least 50\% of all unique timebins covered by the
dataset; topics that do not meet this threshold will not be included in the
returned results.}

\item{minGamma}{the minimum share of a topic per document to be considered
when summarizing topic frequencies; topics with smaller shares per
individual document will be ignored when computing topic frequencies. The
default is \code{0.01}, but should be adjusted with view of the number of
topics and the average length of a document. (In an
\code{\link[stm:stm]{stm topic model}} the likelihood that a topic is
generated from a topic is expressed by the value \emph{gamma}.)}

\item{selectTopicsBy}{the selection approach which determines the metric by
which \code{topic_id}s will be sorted to select the \code{topN} topics.
Currently, the following options are supported: \describe{
\item{most_frequent}{\strong{the default}, select topics based on the total
number of documents in which the topic occurs (\strong{NOTE}, that the
document count depends on the minimum topic likelihood \code{minGamma} that
was specified when obtaining the topic frequencies.)}
\item{trending_up}{select topics with largest upwards trend; internally
this is measured by the slope of a simple linear regression fit to a
\code{topic_id}'s frequency series.} \item{trending_down}{select topics
with largest downward trend; internally this is measured by the slope of a
simple linear regression fit to a \code{topic_id}'s frequency series.}
\item{trending}{select topics with either largest upward or downward trend;
internally this is measured by the absolute value of the slope of a simple
linear regression fit to a \code{topic_id}s frequency series.}
\item{most_volatile}{select topics with the largest change throughout the
covered time period; internally this is measured by the residual standard
deviation of the linear model fit to a \code{topic_id}'s time frequency
series.} \item{topic_id}{select topics specified by \code{topic_id} in the
function argument \code{selectTopics}.} }}

\item{selectTopics}{a vector of topic IDs which should be plotted; this
option is only considered when the option \emph{"topic_id"} is chosen for
\code{selectBy}.}

\item{verboseLabels}{a Boolean indicating if additional topic information
should be used to labels subplots. The \strong{default is \code{FALSE}}.}

\item{nCols}{the number of columns along which topic subplots should be layed
out.}
}
\description{
\code{plot_term_frequencies} plots time series of topic shares for a
selection of topics in a faceted plot. Each topic is displayed in a single
subplot and each time series is overlayed with a linear trendline.
}
\details{
This function merges the computation of topic frequencies
(\code{\link{topic_frequencies}}), creation of suitable labels
(\code{\link{topics_terms_map}}) and the selection of topics by miscellaneous
criteria (\code{\link{select_top_topics}}) into one step.
}
\seealso{
Other visualizations: \code{\link{plot_term_frequencies}}
}
\concept{visualizations}
