% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-analysis.R
\name{primary_topics}
\alias{primary_topics}
\title{Document counts and mean topic shares of the three primary topics for each
document}
\usage{
primary_topics(topicsByDocDate, minGamma = 0)
}
\arguments{
\item{topicsByDocDate}{a dataframe as returned by
\code{\link{topics_by_doc_date}}}

\item{minGamma}{the minimum share of a topic per document to be considered
when summarizing primary topic information; topics with smaller shares per
individual document will be ignored when summarizing the document counts
and mean topic shares. (In an \code{\link[stm:stm]{stm topic model}} the
likelihood that a topic is generated from a topic is expressed by the value
\emph{gamma}.) The default is \code{0}, thus ensuring that three topics are
included for each document.}
}
\value{
a dataframe with 7 columns where: \describe{ \item{topic_id}{a topic
ID as provided as an input in \code{topicsByDocDate}}
\item{n_docs_1}{number of documents where \code{topic_id} has the largest
probability} \item{n_docs_2}{number of documents where \code{topic_id} has
the second largest probability} \item{n_docs_3}{number of documents where
\code{topic_id} has the third largest probability} \item{mean_gamma_1}{mean
probability of all documents in \code{n_docs_1}} \item{mean_gamma_2}{mean
probability of all documents in \code{n_docs_2}} \item{mean_gamma_3}{mean
probability of all documents in \code{n_docs_3}} }
}
\description{
\code{primary_topics} summarizes for each topic the number of documents and
the respective mean topic share (gamma) where a topic is one of the three
primary topics in a document.
}
