% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\docType{data}
\name{sample_dfm}
\alias{sample_dfm}
\title{Document-feature-matrix of the inaugural presidential speeches}
\format{a \code{\link[quanteda:dfm]{quanteda::dfm}}}
\usage{
sample_dfm
}
\description{
This package data provides the document-feature-matrix (DFM) used to fit the
\code{\link[stm:stm]{stm}} topic model \code{\link{sample_topics}}. The DFM
is based on the inaugural presidential speeches in the respective dataset
included in the \emph{quanteda} package
(\code{\link[quanteda:data_corpus_inaugural]{quanteda::data_corpus_inaugural}}),
with stopwords and numbers removed.
}
\details{
This dataset is included for convenience and reproducibility.
}
\keyword{datasets}
