% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-analysis.R
\name{select_top_terms}
\alias{select_top_terms}
\title{Select top terms by count, trend metric and/or name pattern}
\usage{
select_top_terms(termFrequencies, topN = 25,
  selectBy = "most_frequent", selectTerms = NULL)
}
\arguments{
\item{termFrequencies}{a dataframe of \code{term} frequencies as returned by
\code{term_frequencies()}}

\item{topN}{the number of returned top terms meeting the selection criteria
in \code{selectBy}}

\item{selectBy}{the selection approach which determines the metric by which
\code{term}s will be sorted to select the \code{topN} terms. Currently, the
following options are supported: \describe{
\item{most_frequent}{\strong{the default}, select terms based on the total
number of occurrences} \item{trending_up}{select terms with largest upwards
trend; internally this is measured by the slope of a simple linear
regression fit to a \code{term}'s frequency series.}
\item{trending_down}{select terms with largest downward trend; internally
this is measured by the slope of a simple linear regression fit to a
\code{term}'s frequency series.} \item{trending}{select terms with either
largest upward or downward trend; internally this is measured by the
absolute value of the slope of a simple linear regression fit to a
\code{term}s frequency series.} \item{most_volatile}{select terms with the
largest change throughout the covered time period; internally this is
measured by the residual standard deviation of the linear model fit to a
\code{term}'s time frequency series.} }}

\item{selectTerms}{a character vector of term patterns, that terms are
matched to for selection. \code{regular expression} syntax can be applied,
e.g. if \code{c("^mod", "an", "el$", "^outbreak$")} is supplied for
\code{selectTerms}, all terms that \strong{either} start with \emph{'mod'}
\strong{or} contain \emph{'an'} \strong{or} end with \emph{'el'} or the
exact term \emph{'outbreak'} are matched. The arguments \code{selectBy} and
\code{selectTerms} can be combined.}
}
\value{
a dataframe specifying trend metrics employed for selecting top
  \code{term}s, where: \describe{ \item{term}{a unique term}
  \item{n_term_total}{the total number of a \code{term}'s occurrences in the
  dataset} \item{slope}{the slope coefficient of a linear model fit to this
  \code{term}'s time frequency series} \item{volatility}{the residual
  standard deviation of a linear model fit to this \code{term}'s time
  frequency series} \item{trend}{a categorisation of the term frequency
  trend} }
}
\description{
\code{select_top_terms} allows to select a specified number of top terms
based on miscellaneous properties of the term frequencies. This method is
typically used to select term frequency time series for plotting and
exploratory analysis. See the details of the function arguments for selection
options.
}
