% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-analysis.R
\name{select_top_topics}
\alias{select_top_topics}
\title{Select top topics by document counts or temporal trend metric}
\usage{
select_top_topics(topicFrequencies, topN = 25,
  selectBy = "most_frequent", selectTopics = NULL)
}
\arguments{
\item{topicFrequencies}{a dataframe of \emph{topic} frequencies as returned
by \code{\link{topic_frequencies}}}

\item{topN}{the number of returned top topics meeting the selection criteria
in \code{selectBy}}

\item{selectBy}{the selection approach which determines the metric by which
\code{topic_id}s will be sorted to select the \code{topN} terms. Currently,
the following options are supported: \describe{
\item{most_frequent}{\strong{the default}, select terms based on the total
number of documents in which the topic occurs (\strong{NOTE}, that the
document count depends on the minimum topic likelihood \code{minGamma} that
was specified when obtaining the topic frequencies.)}
\item{trending_up}{select topics with largest upwards trend; internally
this is measured by the slope of a simple linear regression fit to a
\code{topic_id}'s frequency series.} \item{trending_down}{select topics
with largest downward trend; internally this is measured by the slope of a
simple linear regression fit to a \code{topic_id}'s frequency series.}
\item{trending}{select topics with either largest upward or downward trend;
internally this is measured by the absolute value of the slope of a simple
linear regression fit to a \code{topic_id}s frequency series.}
\item{most_volatile}{select topics with the largest change throughout the
covered time period; internally this is measured by the residual standard
deviation of the linear model fit to a \code{topic_id}'s time frequency
series.} }}
}
\value{
a dataframe specifying topic metrics employed for selecting top
\code{topic}s, where: \describe{ \item{topic_id}{a unique topic identifier}
\item{n_doc_topics}{the total number of documents in a dataset in which a
topic with \code{topic_id} occurs} \item{slope}{the slope coefficient of a
linear model fit to this \code{topic_id}'s time frequency series}
\item{volatility}{the residual standard deviation of a linear model fit to
this \code{topic_id}'s time frequency series} \item{trend}{a categorisation
of the topic frequency trend} }
}
\description{
\code{select_top_topics} allows to select a specified number of top topics
based on miscellaneous properties of the topic frequencies. This method is
typically used to select a topic frequency time series for plotting and
exploratory analysis. See the details of the function arguments for selection
options.
}
