% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-analysis.R
\name{term_counts}
\alias{term_counts}
\title{Count unique terms in a text source}
\usage{
term_counts(termsByDate)
}
\arguments{
\item{termsByDate}{a dataframe as returned by \code{\link{terms_by_date}}}
}
\value{
a dataframe with three columns listing all unique \code{term}s in the
  provided text source, where \code{n_term} is the total number of occurences
  in the text source and \code{term_share} the share relative to all other
  terms; if no stemming has been applied the term and token in the result are
  identical
}
\description{
\code{term_counts} counts the occurences and computes shares of unique
\code{term}s in a text source. Expects the result of
\code{\link{terms_by_date}} or its variations as input.
}
