% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-analysis.R
\name{term_frequencies}
\alias{term_frequencies}
\title{Compute the frequency of terms within in a chosen time period}
\usage{
term_frequencies(termsByDate, timeBinUnit = "week",
  minTermTimeBins = 0.5, minTermOccurences = 10)
}
\arguments{
\item{termsByDate}{a dataframe as returned by \code{\link{terms_by_date}}}

\item{timeBinUnit}{a character sequence specifying the time period that
should be used as a bin unit when computing term frequencies. Valid values
are \code{"day", "week", "month", "quarter", "year"}, but for the text
sources processed in this package \code{"week"} is recommended and used as
a default. \strong{NOTE:} for the assignment of \code{week}s Monday is
considered as the first day of the week.}

\item{minTermTimeBins}{a double in the range \code{[0,1]} specifying the
minimum share of all unique timebins in which an occurrence of a term must
have been recorded, i.e. a value of \code{0.5} (the default) requires that
an occurrence of a term must have been recorded in at least 50\% of all
unique timebins covered by the dataset; terms that do not meet this
threshold will not be included in the returned results.}

\item{minTermOccurences}{an integer specifying the minimum of total
occurrences of a term to be included in the results; terms that do not meet
this threshold will not be included in the returned results.}
}
\value{
a dataframe with term frequencies by chosen timebin, where:
\describe{ \item{term}{a term as provided as an input in
\code{termsByDate}} \item{timebin}{the first day of the a timebin; if
\code{timeBinUnit} was set to \code{week}, this date will always be a
Monday} \item{n_term_per_timebin}{the number of occurrences of \code{term}
in \code{timebin}} \item{first_occur}{the exact date of the first
occurrence of \code{term} across the whole time range covered by
\code{timebin}s} \item{latest_occur}{the exact date of the latest
occurrence of \code{term} across the whole time range covered by
\code{timebin}s; note that this date can be larger than the maximum
\code{timebin}, as \code{timebin} specifies the floor date of a time unit}
\item{term_share_per_timebin}{the share of \code{term} in a given
\code{timebin} with respect to all other term occurrences in a
\code{timebin}; \strong{NOTE} that this is computed in consideration of all
terms, including those that may be filtered out of the results}
\item{n_term_total}{the total number of occurrences in the dataset, i.e.
across all \code{timebin}s} \item{n_term_timebins}{the number of unique
\code{timebin}s in which an occurrence of \code{term} was recorded} }
}
\description{
\code{term_frequencies} summarizes the counts and relative frequency of terms
in a chosen timebin for a given collection of terms with date of occurrence.
}
\details{
Timebins for which no occurrence of a given term is recorded are added with
an explicit value of zero, excluding however such empty timebins before the
first occurrence of a term and after the last.
}
