% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-analysis.R
\name{terms_tokens_map}
\alias{terms_tokens_map}
\title{Summarize mappings between terms and tokens from a text source}
\usage{
terms_tokens_map(termsByDate)
}
\arguments{
\item{termsByDate}{a dataframe as returned by \code{\link{terms_by_date}}}
}
\value{
a dataframe with four columns:\describe{ \item{term}{a unique
  character sequence identified in a text source} \item{term_instances}{a
  list of unique instances of \code{term} (i.e. \code{token}s) that were
  reduced to \code{term}} \item{n_term_instances}{a list of respective counts
  of the term instances in \code{term_instances}} \item{term_label}{a
  suitable default label for the \code{term}} }
}
\description{
\code{terms_tokens_map} summarizes the result returned by the
\code{\link{terms_by_date}} function(s). The primary purpose is to use this in
labelling results in for example plots.
}
