% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{theme_fte_fonts}
\alias{theme_fte_fonts}
\title{Applies custom fonts and sizes to chart axis}
\usage{
theme_fte_fonts(sizeAxisText = 12)
}
\description{
Applies custom fonts and sizes to chart axis
}
\seealso{
Other visualizations: \code{\link{plot_term_frequencies}},
  \code{\link{theme_stwd_base}}
}
\concept{visualizations}
