% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{theme_stwd_base}
\alias{theme_stwd_base}
\title{Configure a simple plot theme}
\usage{
theme_stwd_base(baseSize = 12, baseFamily = "sans")
}
\arguments{
\item{baseSize}{the font base size. All text elements are sized relative to
this size. Default is \emph{12}.}

\item{baseFamily}{the font family that should be used for text elements.
Default is \emph{sans} (e.g. Helvetica, Arial - depending on which is the
default on the local system).}
}
\description{
\code{theme_stwd_base} provides a default plot theme inspired by the basic
principles explained in Edward Tufte's books as well as plot types showcased
in C. Knaflic's book "Storytelling with data".
}
\details{
It is possible to apply this to any type of plot.
}
\examples{

\dontrun{
theme_stwd_base(baseSize = 12, baseFamily = "serif")
}

}
\seealso{
Other visualizations: \code{\link{plot_term_frequencies}},
  \code{\link{theme_fte_fonts}}
}
\concept{visualizations}
