% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-analysis.R
\name{topic_frequencies}
\alias{topic_frequencies}
\title{Compute topic shares for a given time bin}
\usage{
topic_frequencies(topicsByDocDate, timeBinUnit = "week",
  minGamma = 0.01, minTopicTimeBins = 0.5)
}
\arguments{
\item{topicsByDocDate}{a dataframe as returned by
\code{\link{topics_by_doc_date}}}

\item{timeBinUnit}{a character sequence specifying the time period that
should be used as a bin unit when computing topic share frequencies. Valid
values are \code{"day", "week", "month", "quarter", "year"}, \code{"week"}
is used as a default. \strong{NOTE}, for the assignment of \code{week}s
Monday is considered as the first day of the week.}

\item{minGamma}{the minimum share of a topic per document to be considered
when summarizing topic frequencies, topics with smaller shares per
individual document will be ignored when computing topic frequencies. (In
an \code{\link[stm:stm]{stm topic model}} the likelihood that a topic is
generated from a topic is expressed by the value \emph{gamma}.) The default
is \code{0.01}, but should be adjusted with view of the number of topics
and the average length of a document.}

\item{minTopicTimeBins}{a double in the range \code{[0,1]} specifying the
minimum share of all unique timebins in which an occurrence of a topic
share of at least \code{minGamma} must have been recorded, i.e. a value of
\code{0.5} (the default) requires that an occurrence of a topic must have
been recorded in at least 50\% of all unique timebins covered by the
dataset; topics that do not meet this threshold will not be included in the
returned results.}
}
\value{
a dataframe with term frequencies by chosen timebin, where:
  \describe{ \item{topic_id}{a topic ID as provided as an input in
  \code{topicsByDocDate}} \item{timebin}{the floor date of a timebin; if
  \code{timeBinUnit} was set to \code{week}, this date will always be a
  Monday} \item{median_gamma}{the median of likelihoods of the topic with
  \code{topic_id} in \code{timebin}} \item{mean_gamma}{the mean of
  likelihoods of the topic with \code{topic_id} in \code{timebin}}
  \item{topicshare}{the share of topic with \code{topic_id} relative to all
  topic shares recorded and included in a given \code{timebin}.
  \strong{NOTE:} strictly speaking these are the likelihoods that a document
  is generated from a topic, which we here interpret as the share of a
  document attributed to a topic.} \item{n_docs_topic}{the total number of
  documents in a dataset in which a topic with \code{topic_id} occurs as
  least with likelihood \code{minGamma}} \item{first_occur}{the exact date of
  the first occurrence of a topic with \code{topic_id} across the whole time
  range covered by \code{timebin}s} \item{latest_occur}{the exact date of the
  latest occurrence of a topic with \code{topic_id} across the whole time
  range covered by \code{timebin}s; note that this date can be larger than
  the maximum \code{timebin}, as \code{timebin} specifies the floor date of a
  time unit} \item{n_topic_timebins}{the number of unique \code{timebin}s in
  a topic with \code{topic_id} occurs at least with likelihood
  \code{minGamma}} }
}
\description{
\code{topic_frequencies} summarizes the shares of topics in a chosen time
interval as per provided topic shares by document and date.
}
\details{
A \code{stm} topic model provides for each document the likelihood
(\emph{gamma}) that it is generated from a specific topic; here we interprete
these as the share of a document attributed to this topic and then summarize
these shares per timebin to obtain the share of a topic across all documents
over time.

The topic share or likelihood per document has to be above a threshold
specified by \code{minGamma}. A suitable threshold might consider the number
of topics and the average document size. An additional filtering option is
provided with \code{minTopicTimeBins}.

Timebins for which no occurrence of a given topic is recorded are added with
an explicit value of zero, excluding however such empty timebins before the
first occurrence of a topic and after the last.
}
