% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-extract.R
\name{topics_by_doc_date}
\alias{topics_by_doc_date}
\title{Extract topic likelihoods per document for an STM topic model}
\usage{
topics_by_doc_date(topicModel, termsDfm, textData, documentIdColumn,
  dateColumn)
}
\arguments{
\item{topicModel}{an \code{\link[stm:stm]{stm topic model}} fit to the
document-feature-matrix provided in the \code{termsDfm} argument}

\item{termsDfm}{a \emph{document-feature-matrix} of type
\code{\link[quanteda:dfm]{quanteda::dfm}} that was used to fit the provided
\code{topicModel} (see \code{\link{terms_dfm}}).}

\item{textData}{a dataframe containing the original text source reduced to
the document-feature-matrix provided in the \code{termsDfm} argument}

\item{documentIdColumn}{the column name in \code{textData} specifying a
unique identifier for a document}

\item{dateColumn}{the column name in \code{textData} containing a publication
date associated with a document identified by a unique identifier in
\code{documentIdColumn}}
}
\value{
a dataframe with topic shares/likelihoods per document, where:
  \describe{\item{document}{a unique identifier for a document}
  \item{occur}{a publication date associated with \code{document}}
  \item{topic_id}{a numeric identifier of a topic} \item{gamma}{the
  likelihood that \code{document} is generated from the topic with
  \code{topic_id}; the analyses in this package interprete this likelihood as
  the share of this topic in this document.} }
}
\description{
\code{topics_by_doc_date} retrieves the likelihoods that a document is
generated from a topic (\emph{gamma} in the passed topic STM model). Related
functions and examples in this package interpret this likelihood effectively
as a topic share.
}
