% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-analysis.R
\name{topics_summary}
\alias{topics_summary}
\title{Mean topic likelihoods summary}
\usage{
topics_summary(topicsByDocDate, topicLabels)
}
\arguments{
\item{topicsByDocDate}{a dataframe as returned by
\code{\link{topics_by_doc_date}}}

\item{topicLabels}{a dataframe as returned by \code{\link{topics_terms_map}},
associating a \code{topic_id} with a suitable \code{topic_label}.}
}
\value{
@return a dataframe with term frequencies by chosen timebin, where:
  \describe{ \item{topic_id}{the unique topic identifier assigned by a topic
  model} \item{topic_label}{a character vector of representative labels for
  \code{topic_id}} \item{mean_gamma}{the mean likelihood of \code{topic_id}
  across all documents} }
}
\description{
\code{topics_summary}  summarizes the mean likelihood of topics across all
documents combined with a set of suitable labels. This is a convenience
function to create summary visualizations or interactive tables.
}
