% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-extract.R
\name{topics_terms_map}
\alias{topics_terms_map}
\title{Extract a mapping between topics and representative topic terms from an STM}
\usage{
topics_terms_map(topicModel, nLabelTerms = 7, labelType = "frex")
}
\arguments{
\item{topicModel}{an \code{\link[stm:stm]{stm topic model}}}

\item{nLabelTerms}{the number of terms to be included as a label for a given
topic; default is \code{7}.}

\item{labelType}{the \code{stm} word weighting used to determine the words
most representative of a topic; default is \emph{"frex"} (most frequent and
exclusive terms). Other options are \emph{"prob"}, \emph{"lift"},
\emph{"score"} (see \code{\link[stm:labelTopics]{stm::labelTopics}} for
details)}
}
\value{
a dataframe with topics by ID and a selection of associated topic
terms, where: \describe{ \item{topic_id}{a numeric identifier of a topic}
\item{topic_label}{a character sequence of frequent and exclusive terms
associated with a topic}}
}
\description{
The primary purpose of \code{topics_terms_map} is to provide an association
between topics (by ID) and a representative set of words (terms) for a given
topic. These can be used to create labels for visualizations.
}
\details{
By default the terms that are both most frequent and exclusive are used (see
\code{\link[stm:stm]{stm}} for details).
}
