% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicsplorrr.R
\docType{package}
\name{topicsplorrr}
\alias{topicsplorrr}
\alias{topicsplorrr-package}
\title{topicsplorrr: A Package Supporting Topical Text Analysis, Exploration and
Visualization}
\description{
The package provides functions to analyse and visualize the results of an
[`stm`][stm::stm()] topic model, including a Shiny application for topic
model exploration.
}
